Simplify dependencies between UnifiedEmail and emailcommon
Gradle doesn't make it easy to include individual files. It's easier here to just build
everything in UnifiedEmail and then have emailcommon depend on the output,
rather than try and slice up UnifiedEmail's source into bits which are needed in both.

We're only building these two to get an OtterMail email client, not a platform library